+++ test/Command/ImportCommandTest.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace ShlinkioTest\Shlink\Importer\Command;
 
+use PHPUnit\Framework\Attributes\DataProvider;
+use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 use RuntimeException;
@@ -64,7 +66,7 @@ class ImportCommandTest extends TestCase
         putenv('COLUMNS');
     }
 
-    /** @test */
+    #[Test]
     public function exceptionIsThrownWhenInvalidSourceIsProvided(): void
     {
         $this->expectException(InvalidSourceException::class);
@@ -72,10 +74,7 @@ class ImportCommandTest extends TestCase
         $this->commandTester->execute(['source' => 'invalid']);
     }
 
-    /**
-     * @test
-     * @dataProvider provideSource
-     */
+    #[Test, DataProvider('provideSource')]
     public function dependenciesAreInvokedAsExpected(?string $providedSource, bool $expectSourceQuestion): void
     {
         $source = $providedSource ?? ImportSource::BITLY->value;
@@ -109,16 +108,13 @@ class ImportCommandTest extends TestCase
         }
     }
 
-    public function provideSource(): iterable
+    public static function provideSource(): iterable
     {
         yield 'provided source' => [ImportSource::BITLY->value, false];
         yield 'not provided source' => [null, true];
     }
 
-    /**
-     * @test
-     * @dataProvider provideImportExceptions
-     */
+    #[Test, DataProvider('provideImportExceptions')]
     public function importErrorsAreProperlyHandled(
         ImportException $e,
         int $verbosity,
@@ -148,7 +144,7 @@ class ImportCommandTest extends TestCase
         }
     }
 
-    public function provideImportExceptions(): iterable
+    public static function provideImportExceptions(): iterable
     {
         yield 'no continue token, no verbose' => [
             ImportException::fromError(new RuntimeException('')),