+++ test/Sources/Csv/CsvImporterTest.php
@@ -6,6 +6,8 @@ namespace ShlinkioTest\Shlink\Importer\Sources\Csv;
 
 use DateTimeImmutable;
 use DateTimeInterface;
+use PHPUnit\Framework\Attributes\DataProvider;
+use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
 use Shlinkio\Shlink\Importer\Model\ImportedShlinkUrl;
 use Shlinkio\Shlink\Importer\Sources\Csv\CsvImporter;
@@ -21,13 +23,10 @@ class CsvImporterTest extends TestCase
 
     protected function setUp(): void
     {
-        $this->importer = new CsvImporter($this->getDate());
+        $this->importer = new CsvImporter(self::getDate());
     }
 
-    /**
-     * @test
-     * @dataProvider provideCSVs
-     */
+    #[Test, DataProvider('provideCSVs')]
     public function csvIsProperlyImported(string $csv, string $delimiter, array $expectedList): void
     {
         $options = ImportSource::CSV->toParamsWithCallableMap(
@@ -40,7 +39,7 @@ class CsvImporterTest extends TestCase
         self::assertEquals($expectedList, $urls);
     }
 
-    public function provideCSVs(): iterable
+    public static function provideCSVs(): iterable
     {
         yield 'comma separator' => [
             <<<CSV
@@ -54,7 +53,7 @@ class CsvImporterTest extends TestCase
                     ImportSource::CSV,
                     'https://shlink.io',
                     ['foo', 'bar', 'baz'],
-                    $this->getDate(),
+                    self::getDate(),
                     null,
                     '123',
                     null,
@@ -63,7 +62,7 @@ class CsvImporterTest extends TestCase
                     ImportSource::CSV,
                     'https://facebook.com',
                     [],
-                    $this->getDate(),
+                    self::getDate(),
                     'example.com',
                     '456',
                     'my title',
@@ -83,7 +82,7 @@ class CsvImporterTest extends TestCase
                     ImportSource::CSV,
                     'https://alejandrocelaya.blog',
                     [],
-                    $this->getDate(),
+                    self::getDate(),
                     null,
                     'abc',
                     null,
@@ -92,7 +91,7 @@ class CsvImporterTest extends TestCase
                     ImportSource::CSV,
                     'https://facebook.com',
                     ['foo', 'baz'],
-                    $this->getDate(),
+                    self::getDate(),
                     'example.com',
                     'def',
                     null,
@@ -101,7 +100,7 @@ class CsvImporterTest extends TestCase
                     ImportSource::CSV,
                     'https://shlink.io/documentation',
                     [],
-                    $this->getDate(),
+                    self::getDate(),
                     'example.com',
                     'ghi',
                     'the title',
@@ -122,7 +121,7 @@ class CsvImporterTest extends TestCase
         return $stream; // @phpstan-ignore-line
     }
 
-    private function getDate(): DateTimeInterface
+    private static function getDate(): DateTimeInterface
     {
         static $date;
         return $date ?? ($date = new DateTimeImmutable());