+++ test/Sources/Csv/InvalidPathExceptionTest.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace ShlinkioTest\Shlink\Importer\Sources\Csv;
 
+use PHPUnit\Framework\Attributes\DataProvider;
+use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
 use Shlinkio\Shlink\Importer\Sources\Csv\InvalidPathException;
 
@@ -11,7 +13,7 @@ use function sprintf;
 
 class InvalidPathExceptionTest extends TestCase
 {
-    /** @test */
+    #[Test]
     public function pathNotProvidedCreatesExceptionAsExpected(): void
     {
         $e = InvalidPathException::pathNotProvided();
@@ -19,10 +21,7 @@ class InvalidPathExceptionTest extends TestCase
         self::assertEquals('The path of the file is required.', $e->getMessage());
     }
 
-    /**
-     * @test
-     * @dataProvider providePaths
-     */
+    #[Test, DataProvider('providePaths')]
     public function pathIsNotFileCreatesExceptionAsExpected(string $path): void
     {
         $e = InvalidPathException::pathIsNotFile($path);
@@ -30,7 +29,7 @@ class InvalidPathExceptionTest extends TestCase
         self::assertEquals(sprintf('The file "%s" does not seem to exist. Try another one.', $path), $e->getMessage());
     }
 
-    public function providePaths(): iterable
+    public static function providePaths(): iterable
     {
         yield ['/foo'];
         yield ['/bar'];