+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 
 ### Changed
 * Replaced references to `doma.in` by `s.test`.
+* Migrated to PHPUnit 10.
 
 ### Deprecated
 * *Nothing*